Require contexts in remote client methods
Replace the `context.TODO()`s given to with a context passed in.

Also: the values returned by the docker distribution registry client
will be one of the types in
github.com/docker/distribution/manifest/{schema1,schema2,manifestlist},
so instead of dispatching on the media type and doing the
deserialisation ourselves, just dispatch on the value's type.